chore(deps): update package versions in Cargo.lock
- Bump versions for several dependencies including:
  - agent-client-protocol: 0.9.3 → 0.9.4
  - agent-client-protocol-schema: 0.10.6 → 0.10.8
  - anyhow: 1.0.100 → 1.0.102
  - arc-swap: 1.8.0 → 1.8.2
  - async-executor: 1.13.3 → 1.14.0
  - syn: 2.0.114 → 2.0.117
  - bitflags: 2.10.0 → 2.11.0
  - bumpalo: 3.19.1 → 3.20.2
  - bytes: 1.11.0 → 1.11.1
  - cc: 1.2.55 → 1.2.56
  - chrono: 0.4.43 → 0.4.44
  - deranged: 0.5.5 → 0.5.8

This update ensures compatibility with the latest features and fixes in the respective libraries.

+++ src-tauri/src/adapters/mysql/adapter.rs
@@ -80,7 +80,7 @@ impl MySqlAdapter {
         // This avoids issues with special characters and URL encoding
         let mut builder = OptsBuilder::default()
             .ip_or_hostname(&profile.host)
-            .tcp_port(profile.port as u16)
+            .tcp_port(profile.port)
             .user(Some(&profile.username))
             .db_name(Some(&profile.database));
